pyproject.toml
[tool.poetry]
name = "sleep-analysis"
version = "0.1.0"
description = ""
authors = ["nx1p <email@marci.tech>"]
readme = "README.md"
packages = [{include = "sleep_analysis"}]

[tool.poetry.dependencies]
python = "^3.10"
rich = "^13.7.1"
python-dotenv = "^1.0.1"
pytz = "^2024.1"
requests = "^2.32.3"
psycopg2-binary = "^2.9.9"
aiohttp = "^3.10.9"
asyncpg = "^0.29.0"
py7zr = "^0.22.0"
patool = "^3.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
